Add files via upload
Spatial analysis of tourist areas using social networks is the study and deepening of the concepts and processes of data extraction, data analysis and visualization. The study also contains the use of Machine Learning techniques making a small introduction to the field of data science.
The project presents the technologies used, the theoretical aspects and the implementation of the software project, problems encountered, but also solving or finding alternative solutions to apply the study in a practical format. The site of interest around which the work focused is Airbnb.
All information resulting from the end of the project is attributed to the city of Brașov, Romania. On the Airbnb site it contains about 300 diverse listings. Prices, room type, reviews are just a few criteria applied to this analysis.
 
Finally, the resulting graphs and numbers will show both the average prices of these listings and the most areas of interest in the city.